move headers, replace constexpr to inline
