+++ include/yk/hash/string_hash.hpp
@@ -1,5 +1,5 @@
-#ifndef YK_STRING_HASH_HPP
-#define YK_STRING_HASH_HPP
+#ifndef YK_HASH_STRING_HASH_HPP
+#define YK_HASH_STRING_HASH_HPP
 
 #include <functional>
 #include <string>
@@ -13,14 +13,14 @@ template <class CharT, class Traits = std::char_traits<CharT>>
 struct basic_string_hash {
   using is_transparent = void;
 
-  [[nodiscard]] constexpr std::size_t operator()(const CharT* s) const noexcept { return std::hash<std::basic_string_view<CharT, Traits>>{}(s); }
+  [[nodiscard]] inline std::size_t operator()(const CharT* s) const noexcept { return std::hash<std::basic_string_view<CharT, Traits>>{}(s); }
 
-  [[nodiscard]] constexpr std::size_t operator()(std::basic_string_view<CharT, Traits> sv) const noexcept {
+  [[nodiscard]] inline std::size_t operator()(std::basic_string_view<CharT, Traits> sv) const noexcept {
     return std::hash<std::basic_string_view<CharT, Traits>>{}(sv);
   }
 
   template <class Allocator = std::allocator<CharT>>
-  [[nodiscard]] constexpr std::size_t operator()(const std::basic_string<CharT, Traits, Allocator>& str) const noexcept {
+  [[nodiscard]] inline std::size_t operator()(const std::basic_string<CharT, Traits, Allocator>& str) const noexcept {
     return std::hash<std::basic_string_view<CharT, Traits>>{}(str);
   }
 };
@@ -34,4 +34,4 @@ using u32string_hash = basic_string_hash<char32_t>;
 
 }  // namespace yk
 
-#endif  // YK_STRING_HASH_HPP
+#endif  // YK_HASH_STRING_HASH_HPP
