+++ test/test.cpp
@@ -3,12 +3,12 @@
 #include "yk/hash/adapt.hpp"
 #include "yk/hash/hash_combine.hpp"
 #include "yk/hash/hash_value_for.hpp"
+#include "yk/hash/proxy_hash.hpp"
 #include "yk/hash/range.hpp"
+#include "yk/hash/string_hash.hpp"
 #include "yk/maybe_mutex.hpp"
 #include "yk/par_for_each.hpp"
-#include "yk/proxy_hash.hpp"
 #include "yk/stack.hpp"
-#include "yk/string_hash.hpp"
 #include "yk/util/array_cat.hpp"
 #include "yk/util/auto_sequence.hpp"
 #include "yk/util/forward_like.hpp"
